Moving tests into spec folders
